fix signedness
This commit was SVN r2720.

+++ src/topology-linux.c
@@ -1542,7 +1542,7 @@ hwloc_parse_meminfo_info(struct hwloc_topology *topology,
   while (fgets(string, sizeof(string), fd) && *string != '\0')
     {
       unsigned long long number;
-      if (strlen(string) < prefixlength)
+      if (strlen(string) < (size_t) prefixlength)
         continue;
       if (sscanf(string+prefixlength, "MemTotal: %llu kB", (unsigned long long *) &number) == 1) {
 	*local_memory = number << 10;
